Fix hardware encryption
The qseecomd daemon is necessary to get access to the trusted zone,
which is used to derive the encryption key from the user PIN/password.

The binary and libraries are copied from android_vendor_motorola_surnia,
but should probably be placed somewhere else.  The qseecomd binary is
patched to use /sbin/linker as linker.

Change-Id: Ic8c0ad3afd528f7b1b22f08656fb8d8cd01b17ed